refactor: rename types/types > types and move Tab type to shared
