+++ src/collection/collection.const.ts
@@ -1,4 +1,5 @@
 import { ROUTE_PARTS } from '../shared/constants';
+import { Tab } from '../shared/types';
 
 export const COLLECTION_PATH = Object.freeze({
 	COLLECTION_DETAIL: `/${ROUTE_PARTS.collection}/:id`,
@@ -8,16 +9,16 @@ export const COLLECTION_PATH = Object.freeze({
 // TODO: get these from the api once the database is filled up
 export const USER_GROUPS: string[] = ['Docent', 'Leering', 'VIAA medewerker', 'Uitgever'];
 
-export const COLLECTION_EDIT_TABS = [
+export const COLLECTION_EDIT_TABS: Tab[] = [
 	{
 		id: 'inhoud',
 		label: 'Inhoud',
-		icon: 'collection' as IconName,
+		icon: 'collection',
 	},
 	{
 		id: 'metadata',
 		label: 'Metadata',
-		icon: 'file-text' as IconName,
+		icon: 'file-text',
 	},
 ];
